requirements.txt
streamlit
opencv-python
tensorflow
Pillow
numpy
